refactor: 레이아웃 분리 App.tsx 간소화
